v0.4.1: Progress bar, no more console
Code updates across all four files.  A bunch of new functions / refactoring to get to:

a) Progress bar which updates when loading cards from the current set, and when downloading cards from the set
b) No more print statements!  Everything that is used is in the GUI now.
c) Commented out a bunch of stuff that seemed redundant / deprecated, will completely remove it in another code update.